+++ src/lightly_train/_commands/common_helpers.py
@@ -502,53 +502,3 @@ def _unlink_and_ignore(path: Path) -> None:
         path.unlink(missing_ok=True)
     except OSError:
         pass
-
-
-def get_float32_matmul_precision(
-    float32_matmul_precision: Literal["auto", "highest", "high", "medium"],
-    accelerator: Accelerator,
-    device_ids: list[int],
-) -> Literal["highest", "high", "medium"]:
-    """Get the float32 matmul precision setting."""
-    if float32_matmul_precision == "auto":
-        precision = torch.get_float32_matmul_precision()
-        if isinstance(accelerator, CUDAAccelerator):
-            devices = accelerator.get_parallel_devices(devices=device_ids)
-            if precision == "highest" and all(
-                _device_has_tensor_cores(d) for d in devices
-            ):
-                device_name = torch.cuda.get_device_name(devices[0])
-                # Keep this as debug for now. We should select a better default
-                # precision in the future.
-                logger.debug(
-                    f"You are using a CUDA device with tensor cores ({device_name}). "
-                    "To trade-off precision for performance, set "
-                    "float32_matmul_precision to 'high' or 'medium'. The current "
-                    f"default is '{precision}'. See https://docs.pytorch.org/docs/stable/generated/torch.set_float32_matmul_precision.html#torch.set_float32_matmul_precision "
-                    "for more information."
-                )
-        return precision  # type: ignore[return-value]
-    else:
-        return float32_matmul_precision
-
-
-def _device_has_tensor_cores(device: torch.device) -> bool:
-    """Check if the accelerator has tensor cores."""
-    if torch.cuda.is_available():
-        major, _ = torch.cuda.get_device_capability(device)
-    else:
-        major = 0
-    return device.type == "cuda" and major >= 8
-
-
-@contextlib.contextmanager
-def float32_matmul_precision(
-    float32_matmul_precision: Literal["highest", "high", "medium"],
-) -> Generator[None, None, None]:
-    """Context manager to temporarily set the float32 matmul precision."""
-    current_precision = torch.get_float32_matmul_precision()
-    try:
-        torch.set_float32_matmul_precision(float32_matmul_precision)
-        yield
-    finally:
-        torch.set_float32_matmul_precision(current_precision)
